# getSamples performance improvements