--- a/usb/hantek6022be/data.go
+++ b/usb/hantek6022be/data.go
@@ -55,10 +55,11 @@ const (
 	transferTypeMask  uint8 = 0x03
 	transferTypeIso   uint8 = 0x01
 
-	ch1ID  scope.ChanID = "CH1"
-	ch2ID  scope.ChanID = "CH2"
-	ch1Idx              = 0
-	ch2Idx              = 1
+	ch1ID    scope.ChanID = "CH1"
+	ch2ID    scope.ChanID = "CH2"
+	ch1Idx                = 0
+	ch2Idx                = 1
+	maxChans              = 2
 )
 
 type rateID uint8