Benchmark results:
BenchmarkGetSamples/single_make-4         	   50000	     28906 ns/op
BenchmarkGetSamples/two_makes-4           	   50000	     37339 ns/op
BenchmarkGetSamples/fast_lookup_one_loop-4         	   30000	     50444 ns/op
BenchmarkGetSamples/fast_lookup_two_loops-4        	   50000	     35126 ns/op

PProf results from running a 10 second capture:
before: https://pastebin.com/vv8YBLiT
after: https://pastebin.com/JvxEmdJm

Time spent in getSamples for this run dropped from 17% of the runtime to <5% of the runtime.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/zagrodzki/goscope/40)
<!-- Reviewable:end -->