@dtkaczyk 2017-08-30
Review status: 0 of 2 files reviewed at latest revision, 1 unresolved discussion.

---

*[usb/hantek6022be/capture.go, line 68 at r2](https://reviewable.io:443/reviews/zagrodzki/goscope/40#-KsohKVzLwq91Mk4Kcii:-KsohKW-CBQelz7tCF3k:b-7h96mh) ([raw file](https://github.com/zagrodzki/goscope/blob/cb01b3ff92fac6c467437804127a6f9b1974d94d/usb/hantek6022be/capture.go#L68)):*
> ```Go
> 
> type captureParams struct {
> 	translateSample [2][256]scope.Voltage
> ```

We have a lot of mysterious numbers in the code. It makes it harder to understand what's going on. Maybe we could use constants with meaningful names instead? Not necessarily for all the existing code, but maybe for new commits?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/zagrodzki/goscope/40)*
<!-- Sent from Reviewable.io -->