@zagrodzki 2017-08-30
Review status: 0 of 2 files reviewed at latest revision, 1 unresolved discussion.

---

*[usb/hantek6022be/capture.go, line 68 at r2](https://reviewable.io:443/reviews/zagrodzki/goscope/40#-KsohKVzLwq91Mk4Kcii:-KspLS333OyJLSEmpcVZ:b3jctyo) ([raw file](https://github.com/zagrodzki/goscope/blob/cb01b3ff92fac6c467437804127a6f9b1974d94d/usb/hantek6022be/capture.go#L68)):*
<details><summary><i>Previously, zagrodzki wrote…</i></summary><blockquote>

these are not that mysterious. I can add a const for maxChans = 2 for the size of the outer array.
For the inner array the intention is "all possible byte values". "numBytes" might be confusing, since it typically indicates a length not  number of possible values. maxByte+1 (where maxByte is 255) would be correct, but ugly. Ideas?
</blockquote></details>

I could also just add a comment on translateSample field, that might actually be a better solution? the numbers are not interesting, but the meaning of the entire construct is.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/zagrodzki/goscope/40)*
<!-- Sent from Reviewable.io -->