@zagrodzki 2017-08-31
Review status: 0 of 4 files reviewed at latest revision, 1 unresolved discussion.

---

*[usb/hantek6022be/capture.go, line 68 at r2](https://reviewable.io:443/reviews/zagrodzki/goscope/40#-KsohKVzLwq91Mk4Kcii:-KsrZv552VIFRGztmBSt:bbtp9st) ([raw file](https://github.com/zagrodzki/goscope/blob/cb01b3ff92fac6c467437804127a6f9b1974d94d/usb/hantek6022be/capture.go#L68)):*
<details><summary><i>Previously, dtkaczyk (Dominika Tkaczyk) wrote…</i></summary><blockquote>

Short answer: I would add a comment (it's necessary) and maxChans = 2 (I vote for it, but I'll leave the decision to you). The inner number is not that bad.

Longer explanation: I realize the numbers are not mysterious for the author of them ;). They were mysterious for me. By that I mean: when I look at this declaration, I don't understand what is translateSample, or why its dimensions are what they are. To understand that, I had to analyze another piece of code that is in a different place (the one that uses translateSample). My feeling is that it's not a good situation, I imagine it should be fairly easy to at least have a vague idea what  this piece of code is for without searching in other places.

What bothers me more is that this is just an example, I believe in other files as well there are many similar situations. Of course if you strongly oppose any changes in this practice, say it now and I will remain silent about this issue forever :)
</blockquote></details>

Done.

I'm not opposing any of that, of course. The entire purpose of the peer review is to make sure the code is readable not just to the person who wrote it. You should not need to read multiple places, jumping all around the code, to understand what the code does. Typically the variable or structure declaration should have enough information.

Feel free to point out any other places where you think  the purpose of the code or it's meaning is not clear. It's difficult for me to see that for myself immediately, since everything seems obvious just after you write it...

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/zagrodzki/goscope/40)*
<!-- Sent from Reviewable.io -->